requirements.txt
numpy>=1.21.0
torch>=1.9.0
gym>=0.21.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
pyyaml>=5.4.0
pytest>=6.2.0
seaborn>=0.11.0 